% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOLDprice.R
\name{GOLDprice}
\alias{GOLDprice}
\title{Create Gold price data}
\usage{
GOLDprice(data)
}
\arguments{
\item{data}{dataframes}
}
\value{
A dataframes of gold price data
}
\description{
Create Gold price data
}
