% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monthly_EoP.R
\docType{data}
\name{Monthly_EoP}
\alias{Monthly_EoP}
\title{Monthly_EoP}
\format{
A data frame with 20 variables:
\describe{
\item{Tanggal}{Tanggal}
\item{USD}{Harga emas dalam satuan USD}
\item{EUR}{Harga emas dalam satuan EUR}
\item{JPY}{Harga emas dalam satuan JPY}
\item{GBP}{Harga emas dalam satuan GBP}
\item{CAD}{Harga emas dalam satuan CAD}
\item{CHF}{Harga emas dalam satuan CHF}
\item{INR}{Harga emas dalam satuan INR}
\item{CNY}{Harga emas dalam satuan CNY}
\item{TRY}{Harga emas dalam satuan TRY}
\item{SAR}{Harga emas dalam satuan SAR}
\item{IDR}{Harga emas dalam satuan IDR}
\item{AED}{Harga emas dalam satuan AED}
\item{THB}{Harga emas dalam satuan THB}
\item{VND}{Harga emas dalam satuan VND}
\item{EGP}{Harga emas dalam satuan EGP}
\item{KRW}{Harga emas dalam satuan KRW}
\item{RUB}{Harga emas dalam satuan RUB}
\item{ZAR}{Harga emas dalam satuan ZAR}
\item{AUD}{Harga emas dalam satuan AUD}
}
}
\source{
\describe{
https://www.gold.org/
}
}
\usage{
Monthly_EoP
}
\description{
Monthly_EoP
}
\examples{
data(Monthly_EoP)
}
\keyword{datasets}
